refactoring run.py
I am refactoring so that we can have a new version of fc_run.py for
pbsmrtpipe. It's almost completely cut-and-paste.
